move StreamQueryResult & MemoryQueryResult
