+++ sharding-jdbc/src/main/java/io/shardingsphere/core/executor/result/MemoryQueryResult.java
@@ -15,7 +15,7 @@
  * </p>
  */
 
-package io.shardingsphere.core.executor.type.connection;
+package io.shardingsphere.core.executor.result;
 
 import io.shardingsphere.core.merger.QueryResult;
