+++ sharding-jdbc/src/main/java/io/shardingsphere/core/jdbc/core/statement/ShardingPreparedStatement.java
@@ -25,8 +25,8 @@ import io.shardingsphere.core.constant.SQLType;
 import io.shardingsphere.core.event.ShardingEventBusInstance;
 import io.shardingsphere.core.executor.type.batch.BatchPreparedStatementExecutor;
 import io.shardingsphere.core.executor.type.batch.BatchPreparedStatementUnit;
-import io.shardingsphere.core.executor.type.connection.MemoryQueryResult;
-import io.shardingsphere.core.executor.type.memory.StreamQueryResult;
+import io.shardingsphere.core.executor.result.MemoryQueryResult;
+import io.shardingsphere.core.executor.result.StreamQueryResult;
 import io.shardingsphere.core.executor.type.prepared.PreparedStatementExecutor;
 import io.shardingsphere.core.executor.type.prepared.PreparedStatementUnit;
 import io.shardingsphere.core.jdbc.adapter.AbstractShardingPreparedStatementAdapter;
