+++ sharding-jdbc/src/main/java/io/shardingsphere/core/jdbc/core/statement/ShardingStatement.java
@@ -20,8 +20,8 @@ package io.shardingsphere.core.jdbc.core.statement;
 import com.google.common.base.Optional;
 import io.shardingsphere.core.constant.ConnectionMode;
 import io.shardingsphere.core.constant.SQLType;
-import io.shardingsphere.core.executor.type.connection.MemoryQueryResult;
-import io.shardingsphere.core.executor.type.memory.StreamQueryResult;
+import io.shardingsphere.core.executor.result.MemoryQueryResult;
+import io.shardingsphere.core.executor.result.StreamQueryResult;
 import io.shardingsphere.core.executor.type.statement.StatementExecutor;
 import io.shardingsphere.core.executor.type.statement.StatementUnit;
 import io.shardingsphere.core.jdbc.adapter.AbstractStatementAdapter;
